+++ ReleaseNotes.md
@@ -23,6 +23,8 @@ SOFTWARE.
 ## v3.5
 
 * arm_inlines_ascii: fix lower case result of ASCIIToHex
+* arm_inlines_var_size: fix reducing number of bits in convXtoYBits
+* arm_inlines_float: limit set to 9 digits after floating point in get_fp_dec
 
 ## v3.4